* Preparing to release The Forgotten Server 0.2.1...
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@1784 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ monsters.cpp
@@ -173,7 +173,7 @@ Item* MonsterType::createLootItem(const LootBlock& lootBlock)
 		uint32_t randvalue = Monsters::getLootRandom();
 		if(randvalue < lootBlock.chance)
 		{
-			uint32_t n = randvalue % lootBlock.countmax + 1;
+			uint16_t n = randvalue % lootBlock.countmax + 1;
 			tmpItem = Item::CreateItem(lootBlock.id, n);
 		}
 	}
